Add new model to capture references.
Add command to populate references.

Command will check for:

1. dataset_field.field_dataset
2. dataset_field.field

Matching the dataset editor dataset name. If found it creates a reference
indicating this dataset refers to this dataset.

If either of the above found, the command addtionally checks for
specificaton refering to the referenced by dataset.

If found it will also be set on the reference.

Otherwise no referenes found, nothing set.